requirements.txt
# Core dependencies
python>=3.10
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6

# Social media APIs
tweepy>=4.14.0
facebook-sdk>=3.1.0
google-api-python-client>=2.88.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
snscrape>=0.7.0

# NLP libraries
nltk>=3.8.0
textblob>=0.17.0
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# Data visualization
wordcloud>=1.9.0
plotly>=5.15.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
lxml>=4.9.0